requirements.txt
aenum==3.1.15 ; python_version >= "3.12" and python_version < "4.0"
aiohttp==3.9.3 ; python_version >= "3.12" and python_version < "4.0"
aiosignal==1.3.1 ; python_version >= "3.12" and python_version < "4.0"
annotated-types==0.6.0 ; python_version >= "3.12" and python_version < "4.0"
anyio==4.3.0 ; python_version >= "3.12" and python_version < "4.0"
attrs==23.2.0 ; python_version >= "3.12" and python_version < "4.0"
blinker==1.7.0 ; python_version >= "3.12" and python_version < "4.0"
certifi==2024.2.2 ; python_version >= "3.12" and python_version < "4.0"
charset-normalizer==3.3.2 ; python_version >= "3.12" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.12" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.12" and python_version < "4.0" and platform_system == "Windows"
configparser==6.0.1 ; python_version >= "3.12" and python_version < "4.0"
dataclasses-json==0.6.4 ; python_version >= "3.12" and python_version < "4.0"
deprecated==1.2.14 ; python_version >= "3.12" and python_version < "4.0"
distro==1.9.0 ; python_version >= "3.12" and python_version < "4.0"
flask==3.0.2 ; python_version >= "3.12" and python_version < "4.0"
frozenlist==1.4.1 ; python_version >= "3.12" and python_version < "4.0"
future==1.0.0 ; python_version >= "3.12" and python_version < "4.0"
greenlet==3.0.3 ; python_version >= "3.12" and python_version < "4.0" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32")
grpcio-tools==1.62.1 ; python_version >= "3.12" and python_version < "4.0"
grpcio==1.62.1 ; python_version >= "3.12" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.12" and python_version < "4.0"
h2==4.1.0 ; python_version >= "3.12" and python_version < "4.0"
hpack==4.0.0 ; python_version >= "3.12" and python_version < "4.0"
httpcore==1.0.4 ; python_version >= "3.12" and python_version < "4.0"
httpx==0.27.0 ; python_version >= "3.12" and python_version < "4.0"
httpx[http2]==0.27.0 ; python_version >= "3.12" and python_version < "4.0"
hyperframe==6.0.1 ; python_version >= "3.12" and python_version < "4.0"
idna==3.6 ; python_version >= "3.12" and python_version < "4.0"
itsdangerous==2.1.2 ; python_version >= "3.12" and python_version < "4.0"
jinja2==3.1.3 ; python_version >= "3.12" and python_version < "4.0"
jsonpatch==1.33 ; python_version >= "3.12" and python_version < "4.0"
jsonpointer==2.4 ; python_version >= "3.12" and python_version < "4.0"
langchain-community==0.0.29 ; python_version >= "3.12" and python_version < "4.0"
langchain-core==0.1.33 ; python_version >= "3.12" and python_version < "4.0"
langchain-openai==0.0.8 ; python_version >= "3.12" and python_version < "4.0"
langchain-text-splitters==0.0.1 ; python_version >= "3.12" and python_version < "4.0"
langchain==0.1.13 ; python_version >= "3.12" and python_version < "4.0"
langsmith==0.1.31 ; python_version >= "3.12" and python_version < "4.0"
line-bot-sdk==3.9.0 ; python_version >= "3.12" and python_version < "4.0"
markdown-it-py==3.0.0 ; python_version >= "3.12" and python_version < "4.0"
markupsafe==2.1.5 ; python_version >= "3.12" and python_version < "4.0"
marshmallow==3.21.1 ; python_version >= "3.12" and python_version < "4.0"
mdurl==0.1.2 ; python_version >= "3.12" and python_version < "4.0"
multidict==6.0.5 ; python_version >= "3.12" and python_version < "4.0"
mypy-extensions==1.0.0 ; python_version >= "3.12" and python_version < "4.0"
numpy==1.26.4 ; python_version >= "3.12" and python_version < "4.0"
openai==1.14.3 ; python_version >= "3.12" and python_version < "4.0"
orjson==3.9.15 ; python_version >= "3.12" and python_version < "4.0"
packaging==23.2 ; python_version >= "3.12" and python_version < "4.0"
portalocker==2.8.2 ; python_version >= "3.12" and python_version < "4.0"
protobuf==4.25.3 ; python_version >= "3.12" and python_version < "4.0"
pydantic-core==2.16.3 ; python_version >= "3.12" and python_version < "4.0"
pydantic==2.6.4 ; python_version >= "3.12" and python_version < "4.0"
pygments==2.17.2 ; python_version >= "3.12" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.12" and python_version < "4.0"
pywin32==306 ; python_version >= "3.12" and python_version < "4.0" and platform_system == "Windows"
pyyaml==6.0.1 ; python_version >= "3.12" and python_version < "4.0"
qdrant-client==1.8.0 ; python_version >= "3.12" and python_version < "4.0"
regex==2023.12.25 ; python_version >= "3.12" and python_version < "4.0"
requests==2.31.0 ; python_version >= "3.12" and python_version < "4.0"
rich==13.7.1 ; python_version >= "3.12" and python_version < "4.0"
setuptools==69.2.0 ; python_version >= "3.12" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.12" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.12" and python_version < "4.0"
sqlalchemy==2.0.29 ; python_version >= "3.12" and python_version < "4.0"
tenacity==8.2.3 ; python_version >= "3.12" and python_version < "4.0"
tiktoken==0.6.0 ; python_version >= "3.12" and python_version < "4.0"
tqdm==4.66.2 ; python_version >= "3.12" and python_version < "4.0"
typing-extensions==4.10.0 ; python_version >= "3.12" and python_version < "4.0"
typing-inspect==0.9.0 ; python_version >= "3.12" and python_version < "4.0"
urllib3==2.2.1 ; python_version >= "3.12" and python_version < "4.0"
werkzeug==3.0.1 ; python_version >= "3.12" and python_version < "4.0"
wrapt==1.16.0 ; python_version >= "3.12" and python_version < "4.0"
yarl==1.9.4 ; python_version >= "3.12" and python_version < "4.0"
